Move existing DNC ability logic from scripts to DNC job_utils
